requirements.txt
transformers
torch
sentencepiece
openai-whisper
edge-tts
sounddevice
scipy
numpy

